requirements.txt
llama-index
llama-index-llms-ollama
llama-index-llms-groq
python-dotenv
